NIFI-12318: Fixed byte array generation in GenerateRecord

This closes #8109.

+++ nifi-nar-bundles/nifi-standard-bundle/nifi-standard-processors/src/main/java/org/apache/nifi/processors/standard/GenerateRecord.java
@@ -306,7 +306,7 @@ public class GenerateRecord extends AbstractProcessor {
             case BOOLEAN:
                 return FakerUtils.getFakeData("Bool.bool", faker);
             case BYTE:
-                return faker.number().numberBetween(Byte.MIN_VALUE, Byte.MAX_VALUE);
+                return (byte) faker.number().numberBetween(Byte.MIN_VALUE, Byte.MAX_VALUE);
             case CHAR:
                 return (char) faker.number().numberBetween(Character.MIN_VALUE, Character.MAX_VALUE);
             case DATE:
